fix: CardPropsType을 types폴더 밑에서 관리
